+++ src/lib/config.ts
@@ -1,4 +1,0 @@
-// Only include data that is not sensitive
-export const config = {
-  googleClientId: import.meta.env.VITE_GOOGLE_CLIENT_ID
-}