+++ src/routes/_send-in-blue.ts
@@ -1,11 +1,7 @@
-import dotenv from 'dotenv'
+import { SEND_IN_BLUE_KEY, SEND_IN_BLUE_URL, SEND_IN_BLUE_FROM, SEND_IN_BLUE_ADMINS } from '$env/static/private'
 
-dotenv.config()
-
-const SEND_IN_BLUE_KEY = process.env.SEND_IN_BLUE_KEY
-const SEND_IN_BLUE_URL = process.env.SEND_IN_BLUE_URL
-const SEND_IN_BLUE_FROM = <MessageAddressee> JSON.parse(process.env.SEND_IN_BLUE_FROM || '')
-const SEND_IN_BLUE_ADMINS = <MessageAddressee> JSON.parse(process.env.SEND_IN_BLUE_ADMINS || '')
+const sender = <MessageAddressee> JSON.parse(SEND_IN_BLUE_FROM || '')
+const to  = <MessageAddressee> JSON.parse(SEND_IN_BLUE_ADMINS || '')
 
 // POST or PUT submission to SendInBlue
 const submit = async (method: string, url: string, data: Partial<SendInBlueContact> | SendInBlueMessage) => {
@@ -22,8 +18,5 @@ const submit = async (method: string, url: string, data: Partial<SendInBlueConta
     throw new Error(`Error communicating with SendInBlue.`)
   }
 }
-
-const sender =  SEND_IN_BLUE_FROM
-const to  = SEND_IN_BLUE_ADMINS
 
 export const sendMessage = async (message: Message) => submit('POST', '/v3/smtp/email', { sender, to: [to], ...message })