+++ src/routes/auth/reset/+server.ts
@@ -1,36 +1,35 @@
-import { json as json$1 } from '@sveltejs/kit';
-import dotenv from 'dotenv'
+import { json } from '@sveltejs/kit'
 import type { RequestHandler } from './$types'
-import type  { JwtPayload } from 'jsonwebtoken'
+import type { JwtPayload } from 'jsonwebtoken'
 import jwt from 'jsonwebtoken'
 import { query } from '../../_db'
+import { JWT_SECRET } from '$env/static/private'
 
-dotenv.config()
+export const PUT: RequestHandler = async (event) => {
+	const body = await event.request.json()
+	const { token, password } = body
 
-const JWT_SECRET =  <jwt.Secret> process.env.JWT_SECRET
+	// Check the validity of the token and extract userId
+	try {
+		const decoded = <JwtPayload> jwt.verify(token, <jwt.Secret> JWT_SECRET)
+		const userId = decoded.subject
 
-export const PUT: RequestHandler = async event => {
-  const body = await event.request.json()
-  const { token, password } = body
+		// Update the database with the new password
+		const sql = `CALL reset_password($1, $2);`
+		await query(sql, [userId, password])
 
-  // Check the validity of the token and extract userId
-  try {
-    const decoded = <JwtPayload> jwt.verify(token, JWT_SECRET)
-    const userId = decoded.subject
-
-    // Update the database with the new password
-    const sql = `CALL reset_password($1, $2);`
-    await query(sql, [userId, password])
-
-    return json$1({
-  message: 'Password successfully reset.'
-})
-  } catch (error) {
-    // Technically, I should check error.message to make sure it's not a DB issue
-    return json$1({
-  message: 'Password reset token expired.'
-}, {
-      status: 403
-    })
-  }
+		return json({
+			message: 'Password successfully reset.'
+		})
+	} catch (error) {
+		// Technically, I should check error.message to make sure it's not a DB issue
+		return json(
+			{
+				message: 'Password reset token expired.'
+			},
+			{
+				status: 403
+			}
+		)
+	}
 }